src/main/java/Guitar.java
public class Guitar extends PlayableInstruments implements Playable{

    private int numberOfStrings;

    public Guitar(String instrumentType, String make, String model, String colour, InstrumentFamilyType familyType, double buyPrice, double sellPrice, int quantity, int numberOfStrings) {
        super(instrumentType, make, model, colour, familyType, buyPrice, sellPrice, quantity);
        this.numberOfStrings = numberOfStrings;
    }

    public int getNumberOfStrings() {
        return numberOfStrings;
    }

    @Override
    public String play() {
        return getInstrumentType() + " is being played";
    }
}